+++ IT227_Lab4/202101074_Lab4_1_pow.c
@@ -1,4 +1,4 @@
-long mypow (long a, long b) {
+double mypow (double a, int b) {
     if (b == 0)
         return 1;
     else if (b%2 == 0)
